Veritabani eksikleri giderildi: modeller DbContext'e baglandi, rezervasyon kaydi, migration yapisi, user-secrets, cakisan rezervasyon kontrolu, basari sayfasi, admin rezervasyon/musteri yonetimi, silme onayi

+++ ViewModels/AdminDashboardViewModel.cs
@@ -5,6 +5,6 @@ public class AdminDashboardViewModel
     public int TotalCars { get; set; }
     public int ActiveReservations { get; set; }
     public int Customers { get; set; }
-    public int TotalRevenue { get; set; }
+    public decimal TotalRevenue { get; set; }
     public List<ReservationSummaryViewModel> RecentReservations { get; set; } = new();
 }
